fix: rename contibutions_plotter -> contributions_plotter, cleanup
- Fix filename typo: contibutions_plotter.py -> contributions_plotter.py
- Update all imports to use corrected module name
- Remove unnecessary BaseConfig (Pydantic) inheritance from
  ContributionsPlotter — it's a static-method namespace, not a config model
- Add ValueError validation for invalid 'kind' in plot_contributions()
- Add Raises section to plot_contributions() docstring
